backend/requirements.txt
anyio==3.6.1
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asgiref==3.5.0
asttokens==2.0.8
attrs==22.1.0
autopep8==1.7.0
backcall==0.2.0
bcrypt==4.0.0
beautifulsoup4==4.11.1
bleach==5.0.1
certifi==2022.9.14
cffi==1.15.0
click==8.0.4
cmdstanpy==1.0.7
colorama==0.4.4
contourpy==1.0.5
convertdate==2.4.0
cryptography==36.0.0
cycler==0.11.0
debugpy==1.6.3
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.4
distlib==0.3.6
dj-database-url==0.5.0
Django==3.2.9
django-cors-headers==3.10.0
django-extensions==3.2.0
django-filter==21.1
django-heroku==0.3.1
django-rest-knox==4.1.0
djangorestframework==3.12.4
entrypoints==0.4
ephem==4.1.3
executing==1.1.1
fastjsonschema==2.16.2
filelock==3.8.0
Flask==2.0.3
Flask-Login==0.5.0
Flask-SQLAlchemy==2.5.1
fonttools==4.37.4
greenlet==1.1.2
gunicorn==20.1.0
hijri-converter==2.2.4
holidays==0.16
idna==3.4
ipykernel==6.16.0
ipython==8.5.0
ipython-genutils==0.2.0
ipywidgets==8.0.2
itsdangerous==2.1.0
jedi==0.18.1
Jinja2==3.0.3
joblib==1.2.0
jsonschema==4.16.0
jupyter==1.0.0
jupyter-console==6.4.4
jupyter-core==4.11.1
jupyter-server==1.21.0
jupyter_client==7.4.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.3
kiwisolver==1.4.4
korean-lunar-calendar==0.3.1
LunarCalendar==0.0.9
MarkupSafe==2.1.0
matplotlib==3.6.1
matplotlib-inline==0.1.6
mistune==2.0.4
mysqlclient==2.1.0
nbclassic==0.4.5
nbclient==0.7.0
nbconvert==7.2.1
nbformat==5.7.0
nest-asyncio==1.5.6
notebook==6.5.1
notebook-shim==0.1.0
numpy==1.21.4
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.0.0
pipenv==2022.9.21
platformdirs==2.5.2
prometheus-client==0.15.0
prompt-toolkit==3.0.31
prophet==1.1.1
psutil==5.9.2
psycopg2==2.9.3
pure-eval==0.2.2
pycodestyle==2.9.1
pycparser==2.21
pygame==2.1.2
Pygments==2.13.0
PyMeeus==0.5.11
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-decouple==3.5
pytz==2021.3
pywin32==304
pywinpty==2.0.8
pyzmq==24.0.1
qtconsole==5.3.2
QtPy==2.2.1
scikit-learn==1.0.2
scipy==1.7.3
Send2Trash==1.8.0
setuptools-git==1.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.31
sqlparse==0.4.2
stack-data==0.5.1
terminado==0.16.0
threadpoolctl==3.0.0
tinycss2==1.1.1
toml==0.10.2
tornado==6.2
tqdm==4.64.1
traitlets==5.4.0
tzdata==2022.1
ujson==5.5.0
virtualenv==20.16.5
virtualenv-clone==0.5.7
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.4.1
Werkzeug==2.0.3
whitenoise==6.0.0
widgetsnbextension==4.0.3
